refactor: Update validation logic and exception handling in domain layer
Extracted the minimum password length to a `MIN_LENGTH` constant in `PasswordValidator`.

Updated test cases in `PasswordValidatorTest`, `EmailValidatorTest`, and `NameValidatorTest` to use more specific exception types (`TooShort`, `InvalidFormat`) and removed obsolete tests for password complexity.
