+++ domain/src/commonTest/kotlin/com/moneyplusplus/domain/usecase/PasswordValidator.kt
@@ -3,8 +3,8 @@ package com.moneyplusplus.domain.usecase
 import kotlin.test.Test
 import kotlin.test.assertFailsWith
 import kotlin.test.assertTrue
-import com.moneyplusplus.domain.exception.AppException.ValidationException.Password.Empty
-import com.moneyplusplus.domain.exception.AppException.ValidationException.Password.InvalidPassword
+import com.moneyplusplus.domain.exception.ValidationException.Password.Empty
+import com.moneyplusplus.domain.exception.ValidationException.Password.TooShort
 import com.moneyplusplus.domain.usecase.validation.PasswordValidator
 
 class PasswordValidatorTest {
@@ -20,22 +20,8 @@ class PasswordValidatorTest {
 
     @Test
     fun `password shorter than 8 characters should throw InvalidPassword`() {
-        assertFailsWith<InvalidPassword> {
+        assertFailsWith<TooShort> {
             validator(SHORT_PASSWORD)
-        }
-    }
-
-    @Test
-    fun `password without letters should throw InvalidPassword`() {
-        assertFailsWith<InvalidPassword> {
-            validator(NO_LETTERS_PASSWORD)
-        }
-    }
-
-    @Test
-    fun `password without digits should throw InvalidPassword`() {
-        assertFailsWith<InvalidPassword> {
-            validator(NO_DIGITS_PASSWORD)
         }
     }
 
@@ -48,7 +34,5 @@ class PasswordValidatorTest {
     private companion object {
         const val VALID_PASSWORD = "Abc12345"
         const val SHORT_PASSWORD = "A1b2"
-        const val NO_LETTERS_PASSWORD = "12345678"
-        const val NO_DIGITS_PASSWORD = "abcdefgh"
     }
 }
